# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "guru-core API",
    "description": "\nThe single endpoint the app talks to. A user states a goal; the service collects\nwhatever context exists, asks follow-up questions when something essential is missing,\nand produces three difficulty variants of a scheduled plan.\n\n**Authentication.** Every endpoint except `POST /v1/auth/google`, `GET /health` and the\npresigned `/v1/files/*` routes needs `Authorization: Bearer <jwt>`. Get the token from\n`POST /v1/auth/google`; it carries the user id and expires after `JWT_TTL_SECONDS`.\nThe team-facing `role-models` writes authenticate with `X-API-Key` instead.\n\n**Long-running work.** Plan generation, import parsing and calendar export run on a\nqueue, so the endpoints that start them return `202` with an id. Poll the resource\n(`GET /v1/plan-sessions/{id}`, `GET /v1/imports`) rather than expecting a synchronous\nresult — PostgreSQL is the source of truth for job state, Redis only caches it.\n\n**Errors.** Every failure returns the same envelope:\n`{\"error\": {\"code\": \"not_found\", \"message\": \"...\"}}`. The code is the snake_case name of\nthe domain error: `invalid_input` (422), `unauthorized` (401), `forbidden` (403),\n`not_found` (404), `conflict` (409), `reauth_required` (409), `rate_limited` (429).\n",
    "contact": {
      "name": "Quasar-Gang",
      "url": "https://github.com/Quasar-Gang/guru-core"
    },
    "license": {
      "name": "Proprietary — all rights reserved"
...
  "paths": {
...
    "/v1/auth/google": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange a Google authorization code for an API token",
        "description": "The entry point of the whole API: the only endpoint that needs no bearer token.\n\nThe client runs the Google sign-in flow itself, then posts the one-time `code` here\ntogether with the exact `redirect_uri` it used. The service verifies the code with\nGoogle, finds or creates the user behind that Google account, and returns **our own**\nJWT — a Google token never reaches the client.\n\nSend the returned `access_token` as `Authorization: Bearer <token>` on every other\nendpoint. It expires after `JWT_TTL_SECONDS`; when it does, run this flow again.\n\n`is_new_user` is `true` only on the very first login of an account, which is created\nwith an empty profile and the `UTC` timezone. Use it to decide whether to send the\nuser through onboarding (`PUT /v1/profile`) or straight into the app.\n\nSigning in does **not** grant calendar or spreadsheet access. That is a second,\nseparate consent handled by `GET /v1/integrations/google/authorize`.",
        "operationId": "login_with_google_v1_auth_google_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The API JWT plus the identity behind it, and whether this is a signup.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — Google refused the exchange: the `code` is wrong, already spent or expired, the `redirect_uri` does not match the one the code was issued for, or the returned `id_token` is malformed. Restart the consent flow to obtain a fresh code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`invalid_input` — `code` or `redirect_uri` is missing from the body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests from this client address.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Identify the caller behind the bearer token",
        "description": "Resolve the stored token to a user, without needing to decode the JWT client-side.\n\nUseful on app start to decide between \"signed in\" and \"signed out\": a `200` means the\ntoken is still good, a `401` means it must be replaced via `POST /v1/auth/google`.\nThis is identity only — the questionnaire answers and timezone live on\n`GET /v1/profile`.",
        "operationId": "me_v1_me_get",
        "responses": {
          "200": {
            "description": "The id and email of the user the token belongs to.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — the `Authorization` header is missing, is not a Bearer token, or the JWT is invalid or expired. Sign in again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — the token is valid but the user it points at no longer exists (the account was deleted). Treat it like a signed-out session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ]
      }
    },
    "/v1/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Read the signed-in user's profile",
        "description": "The standing context every plan is generated against: who the user is and where.\n\n`answers` holds the onboarding questionnaire as a free-form JSON object, and\n`timezone` is the IANA zone every scheduled task is placed in. The plan engine reads\nthis on each generation, so keeping it current directly improves plan quality.\n\nA user who has never written a profile is **not** an error: this returns\n`answers: {}`, `timezone: \"UTC\"` and the current time as `updated_at`. There is no\n`404` here — use `is_new_user` from login, or an empty `answers`, to decide whether to\nshow onboarding.",
        "operationId": "get_profile_v1_profile_get",
        "responses": {
          "200": {
            "description": "The stored answers and timezone, or the defaults when nothing has been saved yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — missing, invalid or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Replace the signed-in user's profile",
        "description": "Write the onboarding answers, and later any edit the user makes to them.\n\nThis is a genuine `PUT`: `answers` **replaces** the stored object wholesale, it is not\nmerged key by key. To change one answer, read `GET /v1/profile`, edit the object\nclient-side, and send the whole thing back. Omitting `answers` entirely stores `{}`\nand wipes what was there.\n\n`timezone` is the exception: omit it or send `null` to keep the zone the user already\nhas (or `UTC` on a first write). When present it must be an IANA zone name; it is\nvalidated against the system tz database before anything is saved, so a bad value\nchanges nothing.\n\nTypically called right after a login that returned `is_new_user: true`, and again\nwhenever the user revisits their settings. Later plan generations pick the new values\nup automatically; plans already generated are not rewritten.",
        "operationId": "update_profile_v1_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The profile as stored, including the resolved timezone.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — missing, invalid or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`invalid_input` — `answers` is not a JSON object, one of its keys is not a string, or `timezone` is not a name the IANA database knows (e.g. `Asia/Taipei`, not `GMT+8`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "GoogleLoginRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "redirect_uri": {
            "type": "string",
            "title": "Redirect Uri"
          }
        },
        "type": "object",
        "required": [
          "code",
          "redirect_uri"
        ],
        "title": "GoogleLoginRequest"
      },
...
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "is_new_user": {
            "type": "boolean",
            "title": "Is New User"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user_id",
          "email",
          "is_new_user"
        ],
        "title": "LoginResponse"
      },
...
      "MeResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email"
        ],
        "title": "MeResponse"
      },
...
      "ProfileView": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "answers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Answers"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "answers",
          "timezone",
          "updated_at"
        ],
        "title": "ProfileView"
      },
...
      "UpdateProfileRequest": {
        "properties": {
          "answers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Answers",
            "default": {}
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "type": "object",
        "title": "UpdateProfileRequest"
      },
...
    },
    "securitySchemes": {
      "BearerJWT": {
        "type": "http",
        "description": "A JWT issued by `POST /v1/auth/google`. Send it as `Authorization: Bearer <token>`.",
        "scheme": "bearer"
      }
    }
  },
...
  "servers": [
...
}